64 --> 32
git-svn-id: svn+ssh://frodo/cluster/svn/projects@15367 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda

+++ psc_fsutil_libs/include/psc_util/journal.h
@@ -132,7 +132,7 @@ struct psc_journal {
 #define PJF_ISBLKDEV			(1 << 2)
 
 #define PJE_XID_NONE			0			/* invalid transaction ID */
-#define PJE_MAGIC			UINT64_C(0x4567abcd)
+#define PJE_MAGIC			UINT32_C(0x4567abcd)
 
 /*
  * Journal entry types - lower bits are used internally, higher bits
